online_shop/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.11
django-bootstrap-icons==0.8.7
django-bootstrap5==24.2
django-debug-toolbar==4.4.6
django-phonenumber-field==7.3.0
et-xmlfile==1.1.0
idna==3.7
iniconfig==2.0.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.36
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
